backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
torch
faiss-cpu
numpy
Pillow
cn_clip

